Added back CLI utilities as a separate project + cleanup & attempted bug fix.
Ideally, the CLI utilities would simple import the framework.  I can setup
the CLI project to be dependent on the Git.framework project, and everything
builds, but I can't run the utilities since they can't find the framework..

For now, I simply added all of the files in Git/src into the cli project.
This will at least ensure that both Xcode projects are editing the same set
of files.  Not ideal, but good enough for now.

Re: Failing Tests. All of the Unit Tests were failing with an
Index Out of Bounds error for an [NSString rangeOfString] call.  This was
also causing the CLI utils to crash.  I tracked this down to a method
in GitObjectStore that handles parsing the object type & size.  I added
a fixed for the crash, which causes all of the tests to fail.  Something
about this process is clearly broken, but I'm not sure what is wrong.