Refactor GITPackReverseIndex to be simpler and faster.
Remove the slow, convoluted reverse index creation scheme and replace it with
a CFArray of struct pointers that store the offset and the index of the entry
in the pack index. This is similar to the git implementation, except we use a
CFArray instead of a straight C solution. Note that we need to use a CFArray
instead of NSArray because we want to store pointers to structs, not objects.
NSArray can only store objects.